Refs #45 - Moved search utilities to utils. Added a test
