+++ raspberryio/search/tests/__init__.py
@@ -1,0 +1,1 @@
+from raspberryio.search.tests.test_models import *
